Added Show Employees Button
So far I've just added the "Show Employees" Button which once clicked will only print the names of the employees that are stored in our database. I've also implemented a GET endpoint in the /backend/index.js file, as well as an Axios GET request in the frontend/src/App.js.