Finished The Functionality For Delete Employee
I still have to fix the Popup Modal Component for whenever we choose the "Yes" option such that it disappears from the screen. So far it only disappears when we choose the "No" option. Also, we have to implement an automatic update of the frontend's employees list after we choose to delete an employee, as well as shifting all the id's in the database and reseting the id auto_increment whenever we delete an employee that isn't the last one added via our form.